[#51] Refactor: useFetchCorrections hook 생성
